// src/app/features/models/carrier-org-acct-details-info.model.ts
import { EmailNotificationInfo } from "./email-notification-info";
import { FedExDirectDistInfo } from "./fed-ex-direct-dist-info";
import { FxfrInfo } from "./fxfr-info";
import { ShipExecLabelInfo } from "./ship-exec-label-info";
import { StarTrackInfo } from "./star-track-info";
import { TntInfo } from "./tnt-info";
import { WorldEaseInfo } from "./world-ease-info";

export class CarrierOrgAcctDetailsInfo {
  carrierId:number=0;
  clientId: number=0;
  orgId: string='';
  invOrgId: string='';
  carrierCode: number=0;
  carrierName: string='';
  carrierUrl: string='';
  carrierMode: string='';
  enabledFlag: boolean=false;
  labelStockOrientation: string='';
  docTabLocation: string='';
  userName: string='';
  password: string='';
  labelFormat: string='';
  domain: string='';
  intlLabelFormat: string='';
  eodDocPrinterSymbol: string='';
  eodLabelFormat: string='';
  eodDocStockLabel: string='';
  eodDocPort: string='';
  hubId: string='';
  bolLabelFormat: string='';
  op900LabelFormat: string='';
  intlDocSubType: string='';
  dhlRegionCode: string='';
  docStockSymbol: string='';
  docModelSymbol: string='';
  erpCarrierScacCode: string='';
  erpCarrierName: string='';
  erpCarrierCode: string='';
  defaultSignatureOption: string='';
  genericLabelWithoutRateFlag: boolean=false;
  stampsPaperSize: string='';
  intlLabelOrientation: string='';
  nonDelOption: string='';
  uspsPriceOption: string='';
  uspsReceiptOptions: string='';
  holdForManifest: string='';
  documentPort: string='';
  companyId: string='';
  cn22LabelFormat: string='';
  cn22FormLabelSize: string='';
  siteId: string='';
  starTrackEODUserName: string='';
  starTrackEODPassword: string='';
  starTrackEODAccountNumber: string='';
  starTrackEODUrl: string='';
  dhlLogoFlag: boolean=false;
  dhlLogoImage: any;
  generateSLI: boolean=false;
  customCIDocument: string='';
  defaultRtnShipMethod: string='';
  generateGenericLabel: boolean=false;
  generateSed: boolean=false;
  apiKey: string='';
  labelModelSymbol: string='';
  labelStockSymbol: string='';
  labelPort: string='';
  emailNotificationInfo: EmailNotificationInfo = new EmailNotificationInfo();
  fxfrInfo: FxfrInfo  = new FxfrInfo();
  starTrackInfo: StarTrackInfo = new StarTrackInfo();
  tntInfo: TntInfo = new TntInfo();
  shipExecLabelInfo: ShipExecLabelInfo = new ShipExecLabelInfo();
  worldEaseInfo: WorldEaseInfo = new WorldEaseInfo();
  dhlAPIKey: string='';
  profileName: string='';
  fedexDirectDistInfo:FedExDirectDistInfo = new FedExDirectDistInfo();
  requestOption: string = '';
  shipMethodMappingInfoList: any = [];
}